Thesis/requirements.txt
flask
numpy
pillow
scipy
matplotlib
opencv-python